+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 #Changelog
 
+## v0.4.0 on t2s2-silvio-dev
+- accept empty endDate for slots. It will internally be set to the end of the
+start day but will not be returned via JSON
+
+
 ## v0.3.0 on t2s2-dev
 - use Puma as Webserver on Heroku instead of Unicorn